* valprint.c (print_longest) [CC_HAS_LONG_LONG && PRINTF_HAS_LONG_LONG]: Cast val_long to (long long) or (unsigned long long) to prevent compiler warning on 64-bit systems.

+++ gdb/valprint.c
@@ -332,7 +332,7 @@ print_longest (struct ui_file *stream, int format, int use_local,
       fprintf_filtered (stream,
 			use_local ? local_decimal_format_custom ("ll")
 			: "%lld",
-			val_long);
+			(long long) val_long);
       break;
     case 'u':
       fprintf_filtered (stream, "%llu", (long long) val_long);
@@ -341,13 +341,13 @@ print_longest (struct ui_file *stream, int format, int use_local,
       fprintf_filtered (stream,
 			use_local ? local_hex_format_custom ("ll")
 			: "%llx",
-			val_long);
+			(unsigned long long) val_long);
       break;
     case 'o':
       fprintf_filtered (stream,
 			use_local ? local_octal_format_custom ("ll")
 			: "%llo",
-			val_long);
+			(unsigned long long) val_long);
       break;
     case 'b':
       fprintf_filtered (stream, local_hex_format_custom ("02ll"), val_long);
